feat: review workspace uses Tiptap editor with AI agent
The review page now renders the Tiptap workspace when a proposal is
selected. DReps/SPOs/CC members see the same editor UI as proposers
but in read-only mode with the AI agent in reviewer role.

- Created WorkspaceEmbed: reusable component for editor+agent+status
- Refactored author page to use WorkspaceEmbed (~450 lines removed)
- Review page: replaced ReviewBrief+sidebar with WorkspaceEmbed
- ReviewActionZone preserved as belowEditor slot (voting works)
- Agent uses 'reviewer' or 'cc_member' role based on user segment
- Queue rail unchanged — proposal list still on the left
- WorkspaceLayout accepts className for embedded use (h-full vs h-screen)
